Fix GutterIconCacheTest flakiness
Before it was disabled, testIconUpToDate_entryInvalidSavedChanges
was exhibiting flaky behavior, failing a small percentage of the time.
This flakiness was due to the fact that GutterIconCache was using
VirtualFile#getTimeStamp, which according to the documentation may be
cached.

This CL changes GutterIconCache#isIconUpToDate to use
VirtualFile#getModificationStamp, which gets updated any time the
content of the file is modified. The previously flaky test now
passes consistently and has been re-enabled.

Bug: 64485516
Test: manual confirmation that GutterIcons update correctly, all
  GutterIconTest tests pass, testIconUpToDate_entryInvalidSavedChanges
  passed locally 1000 times in a row

Change-Id: I918c7dc75906b53c3542c8fcaa12032d9b88c588