GradleUpdateTest is causing subsequent tests to skip
GradleUpdateTest is failing in a way that's causing tests
that are run after GradleUpdateTest to skip. By moving
the test to the unreliable group, we can run the test in its
own JVM to observe what is causing the test to fail.

Additionally, there is a minor fix for this test where
the Gradle file might not be updated to use a newer
Gradle plugin until after the background sync task finishes

Test: GradleUpdateTest
Bug: 65559145
Change-Id: I512893f001bab5da887977499f74729bcf679e15