Prevent creation of multiple CpuCaptureView instances
We were creating 2 additional instances of CpuCaptureView when clicking
"Record" if the capture details panel was already opened. The reason was
we were blindly creating an instance when changing profiling status if
the capture was not null, but we shouldn't be creating instances when
transitioning to STARTING or CAPTURING status.

Bug: 65578367
Test: Manual. Verified that no instances of CpuCaptureView is created
      when clicking "Record"

Change-Id: I7e77d9b606545a196dffc241f9222fd179e654a2