Setup global sync listening API
We don't have a build system independent API for globally
listening for any syncs for a given project. Currently the android
module relies on GradleSyncState#subscribe for this functionality.

In order to remove the dependencies on gradle from the android module,
this CL adds the SyncResultListener interface and
PROJECT_SYSTEM_SYNC_TOPIC to AndroidProjectSystem/ProjectSystemUtil.
To listen for sync results for a project, callers should simply
subscribe an instance of SyncResultsListener to the new message
bus topic using MessageBus#connect and MessageBusConnection#subscribe.

Test: updates to GradleSyncStateTest and GradleSyncStateIntegrationTest
Bug: 65626649
Change-Id: I11f85798a3864175a7fcc2ec82b8301d14a25ac0