Fixing condition for when to use patcher
We were incorrectly trying to use the patcher if all of the following:
- not on windows
- had a valid patch from the installed version to the desired version
- Didn't have the patcher installed or it is corrupt
- Wasn't simultaneously trying to install the patcher

In practice this wouldn't happen (except I guess the corrupt patcher
case) since components with patches available always depend on the
patcher, which means that the patcher always must be installed before we
try to install a patch. But that isn't enforced on the studio side, so
we should handle that case.

Test: updated PatchInstallerFactoryTest
Change-Id: Id06b81f996db4c1683e870dc3567c23e4baea603