Migrate usages of GradleSyncState#subscribe
Currently, all classes affected by this CL use GradleSyncState
for global sync status listening. This CL changes these classes
so that they use the build system independent API provided in
AndroidProjectSystem instead.

This is a part of an ongoing effort to move the gradle package
out of the android module and into project-system-gradle.

Test: existing tests for DynamicResourceValueRepository and
      ManifestEditor + manual testing for everything else
Bug: 65626649
Change-Id: I76a868fe7f42ee37b82e8a7daea8bbce3a62c81e