Use build-system-independent check for vector-to-PNG generation
Currently, the vector asset generator checks whether the gradle
plugin in use by the project supports automatic conversion of svg
to raster assets. This CL adds a new API to the project system
that provides this information (whether the PS supports
svg-to-rasterization). The existing check specific to Gradle
has been moved into project-system-gradle.

Test: Manually ran the new vector asset command
Bug: 64976974
Change-Id: I33bf256e11c5d1c43efc31e0c546fe503ebb6847