[AIA] Fix manifest merge for feature modules
When merging the manifest, if the project type is "feature", its
manifest should be merged as an application merge (libraries manifests
should be merged in it) and not as a library merge.

This CL guarantees the manifest is merged correctly for features.

Test: FeatureMergedManifestTest
Fixes bug 65366849

Change-Id: I2fdfdd692bd22308ac61f9c4b95711c77ac3f64e
(cherry picked from commit 403378fdc153116b6e640ccd8495c2171b8bc7b7)