Auto-grant permissions when deploying apk's to Things devices.
When installing an app on a Things device with Instant Run enabled, we
install with parameter -g to auto-grant permissions. Without it, the
developer would have to reboot their device. This is different from a
phone because Things devices generally don't have a screen for the user
to grant permissions.

Test: wrote SplitApkDeployTaskTest
Bug: 65689685
Change-Id: I6409c075c0486cb715f8172c7c8a51263eed7d98