[Bug 65721163] Fix Index out of bounds for empty error messages
When a sync issue is generated, Android Studio uses SyncErrorHandlers
to process such issues. Some of the error handlers assume that the error
message is not empty and try to split the error messages in lines. On
empty messages this list is also empty and all references to any index
will cause an out of bound exception.

This CL makes sure that the list is not empty before trying to access
elements on it.

Test: New test (testEmptyError) to verify all error handlers returned by
SyncErrorHandler.getExtensions() can process an empty error message.
Also enabled testHandleError on SyncErrorHandlerManagerTest

Bugs: 65721163

Change-Id: I5e276933004cc9204482d05953c56a09f0150939
(cherry picked from commit 9376ed8df7ec2a0bbf6c26156526b1809ed01ef0)