refactor AndroidRunConfigurationBase.getState
The method is long (> 150 lines) and handles a lot of logic for Instant
Run.

changes to clean up the method:
- mark variables final if possible
- move code into helper methods with clear names explaining its purpose
- move variable declaration to right before they are used
- clean up code paths as much as possible

While the method is still ~100 lines. The purpose of the lines are more
obvious now:
- if IR enabled and has existing session -> get deploy target from IR
session
- if no target yet, pick target
- if IR enabled, check gradle support
- if debug, check debug
- create LaunchTasksProviderFactory

I would like to write unit tests against AndroidRunConfigurationBase
class but it's tricky due to the various uses of static methods.

Fixes: b/62479087
Test: existing UI tests should cover IR flows, added tests for various
methods in AndroidRunConfigurationBase
Change-Id: I069883657d84615f5f187a422b76ff400f977f93