Remove the explicit "tools:" prefix from completions
For tools attribute we were deliberately adding the tools
namespace prefix in the completion presented to the user.
The completions themselves only contained the attribute
name.
Unfortunately the tools attributes generated by the
Intellij contributor:
   XmlAttributeReferenceCompletionProvider.java
does not include the namespace prefix.
Thus we show some completions with and some without
namespace prefix which looks wrong.

For the android namespace we always omit the namespace
prefix if the completions are requested from a position
after the end of the prefix.
This CL will make the tools attribute work the same way.

Bug: b/66240917

Test: Added unit test that checks the presentable text in
      completion elements.

Change-Id: Ia28a9257b13ea07562cf47529db1f9c88230a9db