Refactor TargetProvider interface
Split the implemtations of creating Targets for
LayoutHandler and its children.

This makes children can have its own TargetProvider
and also uses parent's TargetProvider to create the
ViewGroup specified Targets.
Since parent doesn't need to know it is parent anymore,
this can resolve the case which a component is a parent
and a child in the same time.

Test: Passed SceneTest.
Change-Id: I8653b1adabed130c121536bf472437c54377ce42