Add AndroidProjectSystem.isSyncInProgress()
There's currently no build system-independent way to check
whether or not a sync is currently in progress for a project.
Right now, DelayedInitialization relies on GradleSyncState for
this functionality.

This CL resolves this issue by adding isSyncInProgress to the
project system API and replacing usages of
GradleSyncState#isSyncInProgress with calls to the new method.

Bug: 65626649
Test: trivial redirect to existing method
Change-Id: Iecbdeab95b068073fc92581b946cdd6f817e2892