Fix unreliable NlPreview uitest.
Bug: b/63923598

So it turns out that the production code was tracking the wrong
editor. Which could cause unreliable test results in the tests
that checks the preview for files in split editor windows.

Removed the unreliable group annotation for 2 such tests.
There are 2 additional tests that are marked unreliable in this
test, but I have never seen them fail. Those failures should be
unrelated.

Test: UI test should now be more reliable.

Change-Id: I99a7f6ed74d95ea077657ee34817c4da2916fc4e