Fix testIconUpToDate_entryInvalidDiskChanges
The current implementation of testIconUpToDate_entryInvalidDiskChanges
in GutterIconCacheTest fails when run from a Mac. The test is supposed
to invalidate the icon cache entry for an image resource file by
changing the last modified time of the file to the current system time.

HFS+ (the file system macOS uses) stores timestamps in seconds, and so
it doesn't have fine enough granularity to detect that the file has
been modified when the difference in time is on the order of
milliseconds. This does not cause any errors when running the IDE
(tested manually). It just means that the test gives false negatives
when run on a Mac.

This CL fixes this problem by setting the last modified time to be an
entire second later than the current system time.

Bug: 66469568
Test: test passes consistently on Mac and Linux
Change-Id: I6bb03a7ed61099622e023844086792cdd587838d