Add dependency version retrieval to project system.
This CL adds a build system agnostic way of fetching the version of
a dependency to project system along with supporting classes to
provide an abstraction to dependency management. (e.g. artifact ids,
artifact version, etc)

Along with the addition of getVersion and getDeclaredVersion, the
ImageViewHandler class has been changed by swapping dependsOn with
the new API.

Test: Existing tests pass.
Bug: 65171131
Change-Id: Iedfc39bea00d676e0af3d97e3c70817cae0ad2f3