Fix profiler transform crash on unsupported api < 21.
1.UnsatisfiedLinkError happens for supportlib native code on api 19,
and apk should not be transformed for that. Find api version list
from the device to check. If pre-21 and post-21 are launched together,
transform is not applied.
2.Remove duplicate stop profiling call in setProcess for device switch,
check whether the process to stop is included in all processes from
myDevice. When switching from device above 21 to device below 21,
duplicate stop throws grpc exception originally.

Bug: 37325209
Test: N/A
Change-Id: I60005fd5d90fba92d94b66835363b876e9f2edca