64907322: Show unique artifact names
This CL tweaks the code used to show library names such that
we don't just show the artifact id; we show elements of the
group id too. This is in addition to ag/3007102 which tweaks
the library name method to use fully qualified names, which
is vital when the method is used to produce keys, but not
for display names.

Bug: 64907322
Test: Unit test included
Change-Id: I0f55f0cd06d48ea4912ac4fae9b0fd7a37430234