Dedupe Apply Changes action under Run menu
There were 2 Apply Changes buttons under the Run menu. One is added
automatically from the set of executor buttons on the toolbar declared
in androidstudio.xml.
The 2nd was added in code to follow the Run button.

Since Apply Changes only works on the current configuration
anyways. There shouldn't be an Apply Changes button in the set of
choose config actions.

Removing that and cleaning up how the shortcut is set.

Fixes: 65338131
Test: existing test
Change-Id: Ic3b73c526a32abce45f7253449f328b0049d47db