Switching navigation tests to AndroidTestCase
Previously most navigation tests extended AndroidGradleTestCase,
which meant they could easily and as-in-production include the
navigation library, but also that they took forever to run.
In this change they're almost all switched over to AndroidTestCase,
which is less accurate and somewhat more hacky (since we can't
depend on the navigation aar directly), but runs much faster.

Test: this is it.
Change-Id: Ic16e451c9a9edd4a8c6b199b5ad807263e69c9aa