Attempt to workaround b/67590545 in ui-test.
This ui-test:
  NlPreviewTest.closeSplitLayoutShouldMovePreviewToCorrectFile

seem to be running to an existing bug in Intellij. The problem is
that some events on the event quere are executed during the
tearDown of the test. This CL adds a workaround by simply adding
a delay before the tearDown is executed.
This hack should be removed when the real fix is merged.

Bug: b/63923598

Test: This is a workaround for a ui-test.

Change-Id: I547afff54e6c17f410bdc63b0dba424cc62839ca