Change DevicePicker dialog to respect reuse
Previous if IR is enabled, we ignore the reuse option on the
TargetPickerDialog. A short term fix was hidding the checkbox.

This CL updates the behavior so the IR and non-IR case are consistent.
If IR is enabled and the reuse checkbox is NOT checked, then pressing
"Run" would bring up the dialog to allow user to pick a different deploy
target.

If reuse is checked, then "Run" would auto target the previous device if
it is still online.

Behavior of "Apply Changes" is not changed.

Fixes: b/65638640
Test: updated AndroidRunConfigurationBaseCanInstantRunTest, will likely need to add UI tests
Change-Id: I285aff0784e2aa17cbae9e1b819ffa53abfe0898