project-system: Allow adding dependencies to a module
Currently, we have an API that allows adding a dependency to be made
available to a specific source file. This CL adds a new API that allows
adding dependencies at a higher level - a module. We eventually may
migrate this to add dependency to a variant instead of a module, but
that is pending decisions regarding the builder model.

This CL includes a few example usages of this new API:
LayoutlibCallbackImpl, RenderTask, CreateMultiRootResourceFileAction.
NlDependencyManager has also been retrofitted with the new API, although
it can also be done with the old API provided we don't keep
NlDependencyManager intact.

Test: tests pass.  Added new unit tests to cover NlDependencyManager.  Added FutureUtilsTest to project-system-tests.
Change-Id: Ia89fd77f5fec647f7074b56f80e1e5c02795b2a1