[AVD Mgr] Show "API nn" if the SDK doesn't have a real name
When listing system images for use with a new AVD, AVD Manager
shows the dessert name and API level for each image.

If the API level is new to AVD Manager, it won't know the
dessert name. AVD Manager will then use the <codename> specified
in the image's package.xml file. If that field is absent, AVD
Manager currently displays "null" for the dessert name.

This CL has AVD Manager display "API nn" in that case.

Also, there were some cases where AVD Manager did not update
a displayed icon or text field when a preview images was
selected. This CL ensures that the default icon and preview
name are used in these cases.

Bug: 67694056

Test: Updated ConfigureAvdOptionsStepTest.testUpdateSystemImageData()
      Add     SystemImageListModelTest

Change-Id: I097a0685786887bbbf2bc3e5246101c397578c36