[AVD Mgr] Show "API nn" if the SDK doesn't have a real name
When listing system images for use with a new AVD, AVD Manager
shows the dessert name and API level for each image.

If the API level is new to the AVD Manager, it won't know the
dessert name. AVD Manager will then user the <codename> specified
in the image's package.xml file. If that field is absent, AVD
Manager currently displays "null" for the dessert name.

This CL has the AVD Manager display "API nn" in that case.

Bug: 67694056

Test: Coming soon.

Change-Id: I097a0685786887bbbf2bc3e5246101c397578c36