Add initial support for native code navigation
The current approach uses IntelliJ's Find Usages utility functions to
find matching methods. As a result, we won't have the most accurate
results.

In order to achieve best accuracy, we need to use the search utilities
of cidr-lang module. We can't add the dependency right now because
there is a massive dependency cycle in Android Studio, and making
profilers-android module depend on cidr-lang will cause building issues.

The best approach for now is to introduce a limited code navigation and
improve it once it's safe to depend on cidr-lang module.

Bug: 37294023
Bug: 67844304
Test: Manual. Start/Stop a simpleperf capture. Right-click on native
      methods declared within the project's scope should popup an active
      "Jump to source" menu and clicking on that should open the editor
      on the method's file/line.

Change-Id: I5eac98f471ab10cdf40b8af1afbd58355d0c614d