[AVD Mgr] Improve display of an AVD's disk use
The AVD Manager displays the amount of disk space that an AVD
occupies. This amount was displayed using "MB" or "GB". Because
the values are 1024-based, this CL switches to "MiB" and "GiB".

Also, the numerical value was displayed as an integer truncated
toward zero. This CL rounds the value. It also ensures that there
are at least two significant digits by displaying one fractional
digit for numbers less than 10.

Bug: 65252599

Test: Added AvdDisplayListTest.testStorageSizeDisplayString

Change-Id: I8f3aae4091374e4877b040c5b879f41ad19e49a2