Make DelayedInitialization a project component
DelayedInitialization was marked as a service, but was forcibly
initialized from AndroidGradleProjectComponent, so effectively, it was
activated early at project initialization time.

This CL makes two changes: It registers DelayedInitialization as a
project component, which removes the need to activate it from a
different project component. It also moves the actual work performed to
be done *after* the project has been properly initialized (by moving the
logic to projectOpened()). This paves the way for DelayedInitialization
to be used independent of the build system since the build system in use
may not be available until after a project has been fully initialized.

Test: existing tests + manual test of opening layout editor on new and
existing projects

Change-Id: I62d7ed1acb46eaffe65dabe5adcc14394361811c