[WelcomeWiz] UI Tests for Cancel Wizard
This CL is one of many for replacing the Android Studio
Welcome Wizard (Installation Wizard) from the old Dynamic Wizard to
the new WizardModel framework.
- Added a new UI test to check that if you press "Cancel" during the
"Install Wizard", a dialog is displayed that allows the user 3 choices:
1) Re-run the wizard again on the next start-up.
2) Never show the wizard again.
3) Cancel the "Cancel Dialog" and go back to the Wizard.
- This CL introduces the notion of a WizardDialogFixture and a
StepDialogFixture, and a mechanism for test code to move between Step
and Wizard in a "code chaining" way. This will be later used to
refactor other UI Wizard Tests.
- More tests will be added as the code is being ported.

Test: This is a test
Change-Id: I887ecfad136cb57346a49af2c3519ca5b79ef593