Fix the cut paste id bug on mac
Components' ids were not kept when cutting and pasting a component on the
design surface or component tree on mac.

This also merges the copy paste manager of the component tree and the
DesignSurface. The tests have been updated to reflect this merge and
those which were specific to the copy paste behavior and independent
of the component tree or design surface have been moved to their own
test class.

Bug: 65542961
Test: Test are now passing on mac
Change-Id: Icbd173cc2157f741e0af92c1dfebe47782b47375