Updating action list after add/delete
Also allowing deleting from the list using the delete key.

This also includes a substantial refactor of the way action properties
are done. Previously we created one NlProperty per action in
NavInspectorPanel. Now NavInspectorPanel just checks to see if it's
possible for a selected component to contain actions, and if so creates
a single NavActionsProperty. The NavActionsProperty is then responsible
for creating child properties for each action.
The NavActionListInspectorComponent can now keep a reference to the
parent property, and use it to refresh the list of actions when needed.

Also adopted some Kotlin idioms where I happened to notice them.

Test: Added NavActionsPropertyTest. Updated other existing tests.

Change-Id: I9c80e7a11173bd18cb928abbcfa45c8d7f540108