Add isSyncNeeded and getLastSyncResult to sync APIs
Currently, AndroidModularizeAction, AndroidViewProejctNode,
and DelayedInitialization all make use of GradleSyncState to determine
the result of the latest sync and whether or not any build files have
been updated since the last sync. Since we want to remove explicit
Gradle dependencies from the android module, we need build
system-independent APIs for getting this information.

This CL resolves this issue by adding the isSyncNeeded() and
getLastSyncResult() methods to ProjectSystemSyncManager. It also
introduces SyncResult.UNKNOWN, which getLastSyncResult() will return in
the event that the method cannot determine the result of the last sync
(e.g. if the project has never been synced before).

This CL also creates a separate GradleProjectSystemProvider class and
moves the AndroidProjectSystemProvider implementation logic out of
GradleProjectSystem to prevent against interactions with project
instances that have already been disposed.

GradleSyncState usages in AndoridModularizeAction and
AndroidViewProjectNode have been replaced with calls to
getLastSyncResult(). The usages in DelayedInitialization will be
updated to use the new APIs once we submit change
I62d7ed1acb46eaffe65dabe5adcc14394361811c.

Bug: 65626649, 65213155
Test: GradleProjectSystemTest#testGetLastSyncResult_*
      + manually confirmed that Android view still shows project root
      folder when sync fails
      + manually confirmed that "Modularize" action is grayed-out
      when sync fails

Change-Id: I29f7ef87fc701bc49b8b40f92e9f3c0483e5df24