Update dialog help actions to be consistent.
Dialog help actions should use StudioHelpManagerImpl to link to the
android developer website. This removes the need for each dialog to
track the website url and call browser.

To link to DAC, extend DialogWrapper and override getHelpId and return:
StudioHelpManagerImpl.STUDIO_HELP_PREFIX + url.
StudioHelpManagerImpl will then handle doHelpAction by opening a browser
to DAC with the url append to https://developer.android.com/.

This is the first step in cleaning up help. Ideally we want to switch
studio away from using urls and instead use ids that DAC redirects to
the actual page. This allows DAC to update redirects as the website
changes.

Fixes: 63761943
Test: n/a
Change-Id: I2ca0764645a42983f4be37094f026c87f9f78b9e