Move UI testing code out of AndroidPlugin
All the code in AndroidPlugin that was related to UI testing has been
moved out into a separate application service. UI tests also used
reflection to set a status bit indicating that we are in UI test mode.
This has been cleaned up to be a plugin extension. (This was originally
introduced in I381af280b82b3119984e305b690a2f835920cbdc as a way to not
show the SDK setup wizard).

Test: Both UI tests and non UI test mode work fine
Change-Id: Ic007f10618d1dec389a5145ae735f07d34b4d26d