Fixes various layout/sizing issues with tooltips
This CL fixes an issue where the tooltip would flap
as the length of strings in the tooltips flip between
being long and short.

This CL also fixes some layout issues between the
heading of the tooltip and the legend. It also fixes
an issue with alignment of the legend color boxes
with the heading.

Change-Id: I460835cc0b4f4dc223179b4b33bad18d926a22d1
Fixes: 65156590
Test: Manually verified that the label only increases in size.