Fix bug 67956060: Stabilize JavaToKotlinConversionTest
Formerly, this particular test called select() to
wait for the file to be converted. If the test failed
to find the text to select, then select() would throw
an AssertionError. This error was caught by the
try-catch blocks formerly used in this test.

However, EditorFixture.select() was recently changed
to wait for the cursor position to be at the end of a
matched group. This wait throws an error that did
not get caught by the try-catch blocks, which caused
the test to fail. This test was updated to verify
Java-to-Kotlin conversion differently without
relying on catching AssertionErrors.

Bug: 67956060
Test: JavaToKotlinConversionTest
Change-Id: Ia0bf4aad6f54e66aeb0dd534029f851150fb62b6