MethodNameParser: consider templates and const/volatile
When profiling EndlessTunnel app, the following method can be found:
  Shader::Render(glm::detail::tmat4x4<float, (glm::precision)0>*)

In the previous implementation of MethodNameParser, we were considering
the method's signature was "(glm::precision)0>*)". We need to cover the
case where template arguments also have parentheses. That's fixed by
simply filtering all the templates before extracting the signature.

Another thing we were not covering was the presence of const/volatile
modifiers in the end of the methods, after the signature.  Now, we've
added support to these methods, such as the following:
  art::StackVisitor::GetDexPc(bool) const

Test: MethodNameParserTest
Bug: 37294023
Bug: 67640605
Bug: 68014313
Change-Id: I2ff446d194a9d903fae4d3a7a6a37f102d1ce540