Experiment: run Kotlin in process during integration tests
We seem to be running into sporadic errors of the form:
 Error unmarshaling return header; nested exception is:
  java.io.EOFException
  at sun.rmi.transport.StreamRemoteCall.executeCall(StreamRemoteCall.java:229)
  at sun.rmi.server.UnicastRef.invoke(UnicastRef.java:162)
  at java.rmi.server.RemoteObjectInvocationHandler.invokeRemoteMethod(RemoteObjectInvocationHandler.java:227)
  at java.rmi.server.RemoteObjectInvocationHandler.invoke(RemoteObjectInvocationHandler.java:179)
  at com.sun.proxy.$Proxy117.clearJarCache(Unknown Source)
  at org.jetbrains.kotlin.compilerRunner.GradleCompilerRunner.compileWithDaemon(GradleKotlinCompilerRunner.kt:216)
... both in our post-submit and on go/ab, and one theory is that the
Kotlin compiler daemon is running out of memory and unresponsive/dead
(it is running with -Xmx1024m, not too much).

In this experiment, we run the compiler in process: it has access to
a 4Gb heap, and hopefully we get more information if it crashes.

Test: N/A
Change-Id: I458fe85ed9552bf18198b93018f218a93eaf6b5e