[NPW] No value present when creating a empty activity project
See bug report for full stacktrace
If the SDK Manager can't get the list of SDK's items (for example can't
connect to the network, or some other error), it returns an empty
list of Items. If there is no API levels available for a form factor,
we already display a "Not available" label and disable the form factor
checkbox, but leave it selected (allowing the user to progress to
the next screen).
This CL un-selects the form factor checkbox if there is no API levels
available. If no form factor is selected, that will stop the user from
progressing to the next screen.

Test: N/A can be simulated by returning an empty list
Bug: 68240053
Change-Id: I91dbb6c1691daa4127cdc298ddff471d2c2f4caa