Migrate GradleSyncState usages in DependencyManager
Currently, palette/DependencyManager, palette2/DependencyManager,
and the corresponding tests use GradleSyncState to listen for sync
completions. In order to remove this explicit dependency on Gradle,
this CL replaces usages GradleSyncState with interactions with the
build system-independent PROJECT_SYSTEM_SYNC_TOPIC.

This change is part of an ongoing effort to make the layout editor
build system-independent.

Test: palette/DependencyManagerTest + palette2/DependencyManagerTest
Change-Id: I99612e72251342db488854c2f2ae3c7c833a6be0