Select simpleperf as profiling option for native projects
When we open the CPU profilers, we automatically select a configuration
on the CPU profiling configurations combobox. In the general case, we
should select ART Sampled (sampling java code), but we can make an
educated guess to select simpleperf in case we think the user is
profiling a native project.

The current approach to guess if the project being profiled is native is
checking the files that are currently open and being displayed (i.e. their
tab is active) in Android Studio and checking if at least one of them is
a native file (.c, .cc, .cpp, .h, .hpp). This approach can be changed later
if, for instance,  we get feedback saying it's too aggressive. In that case,
we can check if all (or half) the files being displayed are native.

Test: Added a unit test to CpuProfilerStageTest
Bug: 68014319
Change-Id: Ib01aa6473b9641967b6eb393421961272919eb43