Change the lifetime of InspectorProviders.
The InspectorProviders is a class that holds a reference to all
possible the InspectorProvider registered for the layout
properties. Before this CL this was recreated everytime the
properties were updated which is not optimal and also defeated the
caching done in each InspectorProvider.

With this CL the lifetime becomes the same as the lifetime of the
designer for a xml layout file.

Test: Existing tests.
Change-Id: I082a533dd61e15cb688535518413bad81f7c510d