Rewrite of the SQL resolution logic
Replace the naive implementation with one that tries to follow SQL
scoping rules. The new code follows the recommended approach of
processing the PSI using a Processor that either resolves a reference or
gathers all available variants. Because of SQL structure the walking
part is more complicated than the usual PsiScopeUtil.treeWalkUp, that's
the only tricky bit really. New code supports aliases, subqueries and
joins.

This change is way bigger than I would like, but changes to the
generated PSI classes don't need to be reviewed. There was no way to
break this into smaller pieces without breaking existing tests.

Test: Existing tests pass as before, new tests added for newly enabled
      functionality.
Change-Id: I86977ba79fa31df5772045c093ef8e6f5cb5167a