Don't access device properties from EDT while rendering device name
DeviceRenderer reads some properties of the device to render the device
name. It does this on the UI thread, which means that it blocks the UI
if adb is hung for instance. This CL changes DeviceRender to take in a
list of properties in the form of a DeviceNameProperties. This
properties corresponding to a device are available from a
DeviceNamePropertiesMap. This map takes over the job of retrieving
properties from the device, and keeping an up-to-date copy of properties
for a given device.
When the properties become available, it can invoke a callback which
typically is in charge of refreshing the display and re-rendering based
on the updated data. The DeviceNamePropertiesMap takes care of clearing
the cache when a device is disconnected, and unregistering itself from
listening to the debug bridge when it is disposed.

Test: unit tests added
Bug:67307130
Change-Id: Ia5644726498469d98e65f87f3d9857b4ce2616bd