Use ProjectResourceRepository to lookup styles references
- ThemeHelper was using LocalResourceRepository for looking up
styles which works as long as the current module includes those
styles but does not work in the case where a dependent module includes
the style.
In case of instant apps for example, `feature` depends on `base` which
declares AppTheme to be Theme.AppCompat.
Since ThemeHelper looks at the local repository of `feature` to determine
whether it has an AppCompat theme, it deduces incorrectly that feature
does not use AppCompat.

Bug: 67930622
Test: TemplateTest + Manual tests

Change-Id: Ibe12a669c54a21561675730261572fc65c0d27d4