Wait for instant run APK deployment to emulator to complete
When the test does not wait for the deployment to complete,
the project will close before deployment is completed. This
causes the IDE to show an error dialog, but this dialog
also gets dismissed as the project window closes. This
causes the IDE to store the error. Subsequent tests check
if there are IDE errors and skip if there are any IDE
errors. Thus, subsequent tests were skipping due to
this particular test.

Test: InstantRunTest - all tests should run if this is fixed
Change-Id: I29bf8c55e67510b807e7ddbfcba6d0fe5806efed