Adding Trebuchet to Studio Profilers
The layout of the profilers-atrace folder mirrors the layout
of trebuchet. This will make integrations between the two
branches easier.

A thing to note is any changes we make to trebuchet should
be commented, and we should only make changes that we feel
comfortable reverse integrating to the trebuchet branch.

Test: None
Change-Id: Ib2a6ddc939f1b6ad249645e86292702be63413ba