Cause Gradle Console tool window to close when trying to close console.
If you press Command-W (in Mac, or Ctrl + F4 on Windows) while the Gradle
console is focused, the console will be closed and there is no way to have
it back unless the project is closed and reopend (there is an instance per
project).

With this CL, the content of this tool window is not closable, causing the
window to be closed instead.

Bug: 68304429

Test: New IdeadTestCase test to verify console is not closable.

Change-Id: Ide3a7221e3b5f9594dbf68ff2cf9382914e1cdfb