Add Gradle target type if Gradle facet is present.
This change makes sure the check to add the Android Gradle build target
type is consistent with the check to disable intellijs JavaBuilder. By
doing this we ensure thats it is not possible for both intellijs and android
studios build process is skipped for Gradle projects.

As a result of this, non android Gradle projects no longer skip the
build process.

Test: N/A
Bug: 68749581
Change-Id: Iaf563d3169c63741fecf5f031d5677373e3a57ee