Handle null nodes when getting possible CPU filters
We were not handling the case of null nodes in CaptureModel (e.g. when
the capture itself is null or if the current capture doesn't have a tree
corresponding to the selected thread) when getting the possible filters.
That was causing the capture details panel not to show in some scenarios
(see the bug below).

Test: added unit test
Change-Id: I48d9fe7e7faf5223d1e7d92877a08ed938204965
Fixes: 68932497