Override the internal representation for certain gravity values.
In the tabular view of the properties we display the presence of
a flag value depending on the mapped integer value specified in
attrs.xml.

For gravity the center values are actually zero and the lower bit
is used to indicate that the "axis" is present (see Gravity.java).
This causes e.g. center_vertical to be checked when the user checks
top and bottom which is not ideal.
In addition center_vertical is the default value for a multiline
EditText control. That makes it impossible to remove the check
mark next to center_vertical.

Fix this by internally in the property table handle the top, bottom,
left, right, start, end values as if the "axis" present was not
set for these values.

Bug: b/68335041
Test: Added unit test in NlFlagPropertyItemTest
Change-Id: I64f913e6177ea4da4c97351b64ffdff183d0a7d2