Support adding and removing Layer in DesignSurface
Make DesignSurface can dynamically adding and removing the
Layers. Moreover, move the hoster of Layer from DesignSurface
to SceneView and make the creation of Layer is not lazy. This
is for removing correct Layer when SceneView is not used anymore.

Disposing of Layer is moved from DesignSurface to SceneManager.
For DesignSurface, the Layers of SceneViews are used for
rendering only.

This is the preparation of supporting multiple NlModels.

Test: Updated and passed NavDesignSurfaceTests.
      Updated and passed NlDesignSurfaceTests.

Change-Id: I0245e0300a846223afd41fb33efde46f673c1569