Fix execution workflow for project import error handling.
This CL brings consistency between the error reporting executed
during the first post-import sync and the subsequent regular
syncs. Before, either the project initialisation was not handled
properly, or error processing & quickfixes enrichment would not
happen because external system util would have already submitted
the messages to the UI before enrichment would happen.

Also this CL ensures that we do not report 'Consult IDE log for details'
directly in the notifications pane - this is not what a regular
sync does, and that message will be appended in the gradle console
output anyway.

Notably, a parallel fix for new sync is not needed, because the new
sync infrastructure completely owns error reporting, without any
extra hops through IDEA internals, so runWhenProjectIsInitialised()
is completely fine there.

Fixes: b/68246529, b/66911744
Test: Covered by testHandleErrorWhenProjectNotInitialized for new sync;
also verified manually both for old and new sync.

Change-Id: I9f4f7313f8198ab6b411185e06bcaac80fc1e23b