Fix Java to Kotlin conversion with context menu
The gradle project sync tasks queued in TransactionGuard seem to get
triggered when calling `runProcessWithProgressSynchronously` from
Java to Kotlin converter i.e: JavaToKotlinAction$Companion.convertFiles(..)
This doesn't happen in 3.0 preview 2 when using 1.1.2-4 but happens
with 1.1.3 onwards in 3.0 rc and continues to happen with 1.1.51.

The fix here is to convert the files in response to the rootsChanged()
callback which is triggered after the project has synced. This works both in
the case of a NewProject as well as adding an existing project in 3.1

Bug: 63039711
Test: Manual Test

Change-Id: Id8ba5bc14277a038a7ae657e27de93d795d7ef20