[NPW] Allow add-ons to show in the list of SDKs targets
Currently addons do not work in android studio, they have been broken
since 2.0.
Currently the name of the addon is indistinguishable from other
android SDKs, and the addon path is calculated incorrectly, and finally
the addon doesn't actually function in android studio.
Note: Upgrading an old add-on from Android Studio 1.5, generates an
incorret package.xml in the add-on directory. This CL does not
address that, it just assumes that a valid package.xml is used.

See this issue for more discussion on the addon:
https://code.google.com/p/android/issues/detail?id=199518

Bug: 68038041
Test: Updated AndroidVersionsInfoTest
Change-Id: Ib0791cb6cde6d6acd364f48b8848e77c7923ede1