There is no Psi file for models with an invalid Virtual file
Deleting the layout file of the layout currently displayed in the
Layout Editor would ask for the Psi file which does not exist anymore.
Check if the VirtualFile is invalid instead.
This was broken by commit 39bb590e5c, which switched getFile() from returning
a stored Psi file reference to recomputing the Psi file every time.

Since introducing a getVirtualFile method in NlModel, refactor all
usages of model.getFile().getVirtualFile() to model.getVirtualFile().

Test: NlPreviewTest.deletePreviewedFile
Change-Id: Ia508ad9add0de405711c053a933ce4350cc241ea