Adding capture override for thread state data
This change parses the thread state from the atrace capture and sets the data
on the CpuCapture object. When this is set on the stage the thread
state model is updated for each thread found in the capture.

The model then uses this data as an override in place of returning
previously captured thread state information for the duration of
the capture.

Test: ThreadStateDataSeriesTest, AtraceParserTest, CaptureModelTest, CpuCaptureTest
Change-Id: Ia17c35a1ca05fbfd8ea40e57db9114b7d2adce06